Add project files for Getting Started with Angular2
This project has work demo for tutorials on getting started with Angular2
It is written using Typescript as a language selection
It demos the use of templates both in-line and as external urls
It demos how Angular is modular and uses modules which have components to relate to each other and how those components can be used as a directive.
It demos the different types of binding: Interpolation, property binding, event binding and 2-way binding.
Adding logic using structural directives such as *ngIf and *ngFor
Also data formating with pipes