Adds more features in relation to components
Due to using Typescript which is a very strongly typed language, we introduce Interfaces for those properties or vars that do not have a specific data type.
Add encapsulation of component styles both inline and as an array of linked files.
Demo of life-cycle hooks such as OnInit, OnChange as Interfaces.
Building custom pipes. Here demo'd as the search field.
Using module Id as an alternative to relative paths.