Adds Star Rating as a Nested Component
This demos how to build nested components. Here we use the star rating as a nested component of the product list component (container component). Also how to use the nested within the container.
It also shows how data can be passed from the container to the nested using the @Input()
decorator.
Finally, it also demos how data can be passed from the nested to the container (the other way round) using the @Output() decorator, which can be used on with events.